LayoutTests/platform/ios-simulator/editing/input/style-change-during-input-expected.txt


The text field should accept typed characters.
FAIL window.target.value should be hello. Was .
PASS successfullyParsed is true

TEST COMPLETE
